Replaced the help command with the new help menu powered by reactions
